migration deploy nova

+++ database/migrations/2023_12_24_151641_add_deploy_nova.php
@@ -13,22 +13,22 @@ class AddDeployNova extends Migration
      */
     public function up()
     {
-        // Bảng apartment_asset_adapter đổi cột percent từ integer -> double
+        // Bảng apartment_asset_adapter đổi cột percent từ integer -> float
         if (Schema::hasColumn('apartment_asset_adapter', 'percent')) {
             Schema::table('apartment_asset_adapter', function (Blueprint $table) {
-				$table->double('percent')->change();
+				$table->float('percent')->change();
             });
         }
-        // Bảng apartment_asset_adapter thêm cột change_negotiated_price loại double
+        // Bảng apartment_asset_adapter thêm cột change_negotiated_price loại float
         if (!Schema::hasColumn('apartment_asset_adapter', 'change_negotiated_price')) {
             Schema::table('apartment_asset_adapter', function (Blueprint $table) {
-				$table->double('change_negotiated_price')->nullable();
+				$table->float('change_negotiated_price')->nullable();
             });
         }
-        // Bảng apartment_asset_comparison_factors thêm cột adjust_coefficient dạng double default 100
+        // Bảng apartment_asset_comparison_factors thêm cột adjust_coefficient dạng float default 100
         if (!Schema::hasColumn('apartment_asset_comparison_factors', 'adjust_coefficient')) {
             Schema::table('apartment_asset_comparison_factors', function (Blueprint $table) {
-				$table->double('adjust_coefficient')->nullable()->default(100);
+				$table->float('adjust_coefficient')->nullable()->default(100);
             });
         }
         // Bảng apartment_asset_properties thêm cột loai_can_ho_id dạng integer
@@ -37,34 +37,34 @@ class AddDeployNova extends Migration
 				$table->integer('loai_can_ho_id')->nullable();
             });
         }
-        // Bảng appraise_adapter thêm cột change_negotiated_price dạng double
+        // Bảng appraise_adapter thêm cột change_negotiated_price dạng float
         if (!Schema::hasColumn('appraise_adapter', 'change_negotiated_price')) {
             Schema::table('appraise_adapter', function (Blueprint $table) {
-				$table->double('change_negotiated_price')->nullable();
+				$table->float('change_negotiated_price')->nullable();
             });
         }
-        // Bảng appraise_comparison_factor thêm cột adjust_coefficient dạng double default 100
+        // Bảng appraise_comparison_factor thêm cột adjust_coefficient dạng float default 100
         if (!Schema::hasColumn('appraise_comparison_factor', 'adjust_coefficient')) {
             Schema::table('appraise_comparison_factor', function (Blueprint $table) {
-				$table->double('adjust_coefficient')->nullable()->default(100);
+				$table->float('adjust_coefficient')->nullable()->default(100);
             });
         }
-        // Bảng certificate_apartment_adapters đổi cột percent từ integer -> double
+        // Bảng certificate_apartment_adapters đổi cột percent từ integer -> float
         if (Schema::hasColumn('certificate_apartment_adapters', 'percent')) {
             Schema::table('certificate_apartment_adapters', function (Blueprint $table) {
-				$table->double('percent')->change();
+				$table->float('percent')->change();
             });
         }
-        // Bảng certificate_apartment_adapters thêm cột change_negotiated_price loại double
+        // Bảng certificate_apartment_adapters thêm cột change_negotiated_price loại float
         if (!Schema::hasColumn('certificate_apartment_adapters', 'change_negotiated_price')) {
             Schema::table('certificate_apartment_adapters', function (Blueprint $table) {
-				$table->double('change_negotiated_price')->nullable();
+				$table->float('change_negotiated_price')->nullable();
             });
         }
-        // Bảng certificate_apartment_comparison_factors thêm cột adjust_coefficient dạng double default 100
+        // Bảng certificate_apartment_comparison_factors thêm cột adjust_coefficient dạng float default 100
         if (!Schema::hasColumn('certificate_apartment_comparison_factors', 'adjust_coefficient')) {
             Schema::table('certificate_apartment_comparison_factors', function (Blueprint $table) {
-				$table->double('adjust_coefficient')->nullable()->default(100);
+				$table->float('adjust_coefficient')->nullable()->default(100);
             });
         }
         // Bảng certificate_apartment_properties thêm cột loai_can_ho_id dạng integer
@@ -73,22 +73,22 @@ class AddDeployNova extends Migration
 				$table->integer('loai_can_ho_id')->nullable();
             });
         }
-        // Bảng certificate_asset_adapter đổi cột percent từ integer -> double
+        // Bảng certificate_asset_adapter đổi cột percent từ integer -> float
         if (Schema::hasColumn('certificate_asset_adapter', 'percent')) {
             Schema::table('certificate_asset_adapter', function (Blueprint $table) {
-				$table->double('percent')->change();
+				$table->float('percent')->change();
             });
         }
-        // Bảng certificate_asset_adapter thêm cột change_negotiated_price loại double
+        // Bảng certificate_asset_adapter thêm cột change_negotiated_price loại float
         if (!Schema::hasColumn('certificate_asset_adapter', 'change_negotiated_price')) {
             Schema::table('certificate_asset_adapter', function (Blueprint $table) {
-				$table->double('change_negotiated_price')->nullable();
+				$table->float('change_negotiated_price')->nullable();
             });
         }
-        // Bảng certificate_asset_comparison_factor thêm cột adjust_coefficient dạng double default 100
+        // Bảng certificate_asset_comparison_factor thêm cột adjust_coefficient dạng float default 100
         if (!Schema::hasColumn('certificate_asset_comparison_factor', 'adjust_coefficient')) {
             Schema::table('certificate_asset_comparison_factor', function (Blueprint $table) {
-				$table->double('adjust_coefficient')->nullable()->default(100);
+				$table->float('adjust_coefficient')->nullable()->default(100);
             });
         }
         // Bảng room_details thêm cột loai_can_ho_id dạng integer
@@ -109,19 +109,19 @@ class AddDeployNova extends Migration
      */
     public function down()
     {
-        // Bảng apartment_asset_adapter đổi cột percent từ integer -> double (reverse)
+        // Bảng apartment_asset_adapter đổi cột percent từ integer -> float (reverse)
         if (Schema::hasColumn('apartment_asset_adapter', 'percent')) {
             Schema::table('apartment_asset_adapter', function (Blueprint $table) {
 				$table->integer('percent')->change();
             });
         }
-        // Bảng apartment_asset_adapter thêm cột change_negotiated_price loại double (reverse)
+        // Bảng apartment_asset_adapter thêm cột change_negotiated_price loại float (reverse)
         if (Schema::hasColumn('apartment_asset_adapter', 'change_negotiated_price')) {
             Schema::table('apartment_asset_adapter', function (Blueprint $table) {
 				$table->dropColumn('change_negotiated_price');
             });
         }
-        // Bảng apartment_asset_comparison_factors thêm cột adjust_coefficient dạng double default 100 (reverse)
+        // Bảng apartment_asset_comparison_factors thêm cột adjust_coefficient dạng float default 100 (reverse)
         if (Schema::hasColumn('apartment_asset_comparison_factors', 'adjust_coefficient')) {
             Schema::table('apartment_asset_comparison_factors', function (Blueprint $table) {
 				$table->dropColumn('adjust_coefficient');
@@ -133,31 +133,31 @@ class AddDeployNova extends Migration
 				$table->dropColumn('loai_can_ho_id');
             });
         }
-        // Bảng appraise_adapter thêm cột change_negotiated_price dạng double (reverse)
+        // Bảng appraise_adapter thêm cột change_negotiated_price dạng float (reverse)
         if (Schema::hasColumn('appraise_adapter', 'change_negotiated_price')) {
             Schema::table('appraise_adapter', function (Blueprint $table) {
 				$table->dropColumn('change_negotiated_price');
             });
         }
-        // Bảng appraise_comparison_factor thêm cột adjust_coefficient dạng double default 100 (reverse)
+        // Bảng appraise_comparison_factor thêm cột adjust_coefficient dạng float default 100 (reverse)
         if (Schema::hasColumn('appraise_comparison_factor', 'adjust_coefficient')) {
             Schema::table('appraise_comparison_factor', function (Blueprint $table) {
 				$table->dropColumn('adjust_coefficient');
             });
         }
-        // Bảng certificate_apartment_adapters đổi cột percent từ integer -> double (reverse)
+        // Bảng certificate_apartment_adapters đổi cột percent từ integer -> float (reverse)
         if (Schema::hasColumn('certificate_apartment_adapters', 'percent')) {
             Schema::table('certificate_apartment_adapters', function (Blueprint $table) {
 				$table->integer('percent')->change();
             });
         }
-        // Bảng certificate_apartment_adapters thêm cột change_negotiated_price loại double (reverse)
+        // Bảng certificate_apartment_adapters thêm cột change_negotiated_price loại float (reverse)
         if (Schema::hasColumn('certificate_apartment_adapters', 'change_negotiated_price')) {
             Schema::table('certificate_apartment_adapters', function (Blueprint $table) {
 				$table->dropColumn('change_negotiated_price');
             });
         }
-        // Bảng certificate_apartment_comparison_factors thêm cột adjust_coefficient dạng double default 100 (reverse)
+        // Bảng certificate_apartment_comparison_factors thêm cột adjust_coefficient dạng float default 100 (reverse)
         if (!Schema::hasColumn('certificate_apartment_comparison_factors', 'adjust_coefficient')) {
             Schema::table('certificate_apartment_comparison_factors', function (Blueprint $table) {
 				$table->dropColumn('adjust_coefficient');
@@ -169,19 +169,19 @@ class AddDeployNova extends Migration
 				$table->dropColumn('loai_can_ho_id');
             });
         }
-        // Bảng certificate_asset_adapter đổi cột percent từ integer -> double (reverse)
+        // Bảng certificate_asset_adapter đổi cột percent từ integer -> float (reverse)
         if (Schema::hasColumn('certificate_asset_adapter', 'percent')) {
             Schema::table('certificate_asset_adapter', function (Blueprint $table) {
 				$table->integer('percent')->change();
             });
         }
-        // Bảng certificate_asset_adapter thêm cột change_negotiated_price loại double (reverse)
+        // Bảng certificate_asset_adapter thêm cột change_negotiated_price loại float (reverse)
         if (Schema::hasColumn('certificate_asset_adapter', 'change_negotiated_price')) {
             Schema::table('certificate_asset_adapter', function (Blueprint $table) {
 				$table->dropColumn('change_negotiated_price');
             });
         }
-        // Bảng certificate_asset_comparison_factor thêm cột adjust_coefficient dạng double default 100 (reverse)
+        // Bảng certificate_asset_comparison_factor thêm cột adjust_coefficient dạng float default 100 (reverse)
         if (Schema::hasColumn('certificate_asset_comparison_factor', 'adjust_coefficient')) {
             Schema::table('certificate_asset_comparison_factor', function (Blueprint $table) {
 				$table->dropColumn('adjust_coefficient');
